+++ CHANGELOG.rst
@@ -25,6 +25,8 @@ Added
   EMsoft's HDF5 files. (`#476 <https://github.com/pyxem/kikuchipy/pull/476>`_)
 - `ECPMasterPattern` class to store Electron Channeling Pattern (ECP) master patterns.
   (`#476 <https://github.com/pyxem/kikuchipy/pull/476>`_)
+- `progressbar` parameter to functions downloading external datasets in the data module.
+  (`#515 <https://github.com/pyxem/kikuchipy/pull/515>`_)
 - Support for Python 3.10. (`#504 <https://github.com/pyxem/kikuchipy/pull/504>`_)
 - `EBSD.static_background` property for easier access to the background pattern.
   (`#475 <https://github.com/pyxem/kikuchipy/pull/475>`_)