+++ CHANGELOG.rst
@@ -12,6 +12,23 @@ best to adhere to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 Contributors to each release are listed in alphabetical order by first name. List
 entries are sorted in descending chronological order.
 
+0.5.8 (2022-05-16)
+==================
+
+Contributors
+------------
+- Håkon Wiik Ånes
+
+Changed
+-------
+- Minimal version of ``orix`` is increased to 0.9.
+  (`#520 <https://github.com/pyxem/kikuchipy/pull/520>`_)
+
+Fixed
+-----
+- Internal use of ``orix.vector.Vector3d`` following ``orix``' 0.9.0 release.
+  (`#520 <https://github.com/pyxem/kikuchipy/pull/520>`_)
+
 0.5.7 (2022-01-10)
 ==================